refactor: extract path resolution and demo test data (Phase 2)
- Create NativeLibraryResolver.cs: centralized path resolution for native
  libraries and dictionaries (~340 lines)
- Update OpenJTalkPhonemizer.cs: remove ~395 lines by using NativeLibraryResolver
- Create DemoTestData.cs: extract test phrases and model config from demo
- Update InferenceEngineDemo.cs: reduce ~110 lines by using DemoTestData
- Update IOSBuildValidationTest: check NativeLibraryResolver for iOS platform code

Total reduction: ~500 lines
