+++ Assets/uPiper/Runtime/Core/Platform/NativeLibraryResolver.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4d241faf14a253b4b9eae39f4646a5d2
